// src/overlay/overlay-state-types.ts
export interface OverlayImageInfo {
    naturalWidth: number;
    naturalHeight: number;
    mimeType?: 'image/jpeg' | 'image/png' | 'image/webp';
    orientation?: 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8;
    sourceId?: string;
    checksum?: string;
}

export interface OverlayBaseImageTransform {
    rotation?: number;
    flipX?: boolean;
    flipY?: boolean;
}